pyproject.toml
[tool.poetry]
name = "food marketplace api"
version = "0.1.0"
description = ""
authors = ["BRANYA <tbd2504@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
django-baton = "^3.1.0"
django-cors-headers = "^4.3.1"
django-extensions = "^3.2.3"
django-filter = "^24.2"
django-split-settings = "^1.3.1"
drf-spectacular = "^0.27.2"
djangorestframework = "^3.15.1"
gunicorn = "^22.0.0"
python-dotenv = "^1.0.1"
pillow = "^10.3.0"
djangorestframework-simplejwt = "^5.3.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
ruff = "^0.4.6"
django-debug-toolbar = "^4.4.2"
faker = "^25.3.0"
mypy = "^1.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "single"

[tool.mypy]
ignore_missing_imports = true
no_implicit_optional = false
